majority done

+++ server/utils/deleteFile.js
@@ -1,6 +1,7 @@
 const fs = require('fs')
 
 function deleteFile(filePath){
+    if (!fs.existsSync(filePath)) return;
     fs.unlink(filePath, (err)=>{
         if (err){
             console.error('Failed to delete file')
